test multi-class metric
